+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 
 ### :house: Internal
 
-  - [#2608](https://github.com/wix-incubator/rich-content/pull/2608) moved MobileSettingsHeader/UrlInputModal components into ui-components folder
+  - [#2608](https://github.com/wix-incubator/rich-content/pull/2608) moved SettingsMobileHeader/UrlInputModal components into ui-components folder
 
 </details>
 <hr/>