+++ CHANGELOG.md
@@ -17,6 +17,15 @@
     Click to see more.
   </summary>
 
+### :house: Internal
+
+  - [#2608](https://github.com/wix-incubator/rich-content/pull/2608) moved SettingsMobileHeader/UrlInputModal components into ui-components folder
+
+</details>
+<hr/>
+
+## 8.44.1 (Jun 9, 2021)
+
 ### :bug: Bug Fix
 
 - `editor`
@@ -36,9 +45,6 @@
 
 - `ui-components`
   - [#2607](https://github.com/wix-incubator/rich-content/pull/2607) moved SelectionList/Tabs/Tab/Image/SliderPanel components into ui-components folder
-
-</details>
-<hr/>
 
 ## 8.44.0 (Jun 8, 2021)